got logic in a better state before moving on to more reworking of the structure and code
