+++ epub-lint/internal/ui/common.go
@@ -7,6 +7,7 @@ import (
 	"github.com/charmbracelet/lipgloss"
 	"github.com/charmbracelet/x/ansi"
 	rw "github.com/mattn/go-runewidth"
+	stringdiff "github.com/pjkaufman/go-go-gadgets/pkg/string-diff"
 )
 
 // icons
@@ -115,4 +116,33 @@ func wrap(runes []rune, width int) [][]rune {
 // from https://github.com/charmbracelet/bubbles/blob/d66fddf5e780b2bf30e386dbf4e65b55b258197f/textarea/textarea.go#L1463-L1465
 func repeatSpaces(n int) []rune {
 	return []rune(strings.Repeat(string(' '), n))
+}
+
+func getStringDiff(original, new string) (string, error) {
+	return stringdiff.GetPrettyDiffString(strings.TrimLeft(original, "\n"), strings.TrimLeft(new, "\n"))
+}
+
+// textarea gets rid of tabs when creating changes, so in order to preserve tabs in the starting whitespace of a line
+// we will use the value of original as the template for what whitespace is needed for each line present
+func alignWhitespace(original, new string) string {
+	origLines := strings.Split(original, "\n")
+	newLines := strings.Split(new, "\n")
+
+	var min = len(newLines)
+	if len(origLines) < min {
+		min = len(origLines)
+	}
+
+	for i := 0; i < min; i++ {
+		origPrefix := ""
+		for j := 0; j < len(origLines[i]); j++ {
+			if !unicode.IsSpace(rune(origLines[i][j])) {
+				break
+			}
+			origPrefix += string(origLines[i][j])
+		}
+		newLines[i] = origPrefix + strings.TrimLeft(newLines[i], " \t")
+	}
+
+	return strings.Join(newLines, "\n")
 }
