+++ epub-lint/internal/ui/fixable-issues.go
@@ -2,12 +2,9 @@ package ui
 
 import (
 	"errors"
-	"fmt"
 	"io"
 	"strings"
-	"unicode"
 
-	"github.com/atotto/clipboard"
 	"github.com/charmbracelet/bubbles/textarea"
 	"github.com/charmbracelet/bubbles/textinput"
 	"github.com/charmbracelet/bubbles/viewport"
@@ -15,11 +12,6 @@ import (
 	"github.com/charmbracelet/lipgloss"
 	"github.com/pjkaufman/go-go-gadgets/epub-lint/cmd/tui"
 	potentiallyfixableissue "github.com/pjkaufman/go-go-gadgets/epub-lint/internal/potentially-fixable-issue"
-	stringdiff "github.com/pjkaufman/go-go-gadgets/pkg/string-diff"
-)
-
-var (
-	maxDisplayHeight = 20
 )
 
 var ErrUserKilledProgram = errors.New("user killed program")
@@ -57,6 +49,11 @@ type sectionBreakStageInfo struct {
 	contextBreak *string
 }
 
+// type block struct {
+// 	text  string
+// 	width int
+// }
+
 type PotentiallyFixableStageInfo struct {
 	FilePaths                                                                           []string
 	FileTexts                                                                           []string
@@ -70,8 +67,6 @@ type PotentiallyFixableStageInfo struct {
 	suggestionEdit                                                                      textarea.Model
 	suggestionDisplay                                                                   viewport.Model
 	scrollbar                                                                           tea.Model
-
-	// leftStatusWidth int
 }
 
 type CssSelectionStageInfo struct {
@@ -150,161 +145,47 @@ func (m FixableIssuesModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 		initialStage = m.currentStage
 	)
 
-	m.sectionBreakInfo.input, cmd = m.sectionBreakInfo.input.Update(msg)
-	cmds = append(cmds, cmd)
+	// general logic for handling keys here
+	switch msg := msg.(type) {
+	case tea.KeyMsg:
+		switch msg.String() {
+		case "ctrl+c":
+			m.Err = ErrUserKilledProgram
 
-	m.PotentiallyFixableIssuesInfo.suggestionEdit, cmd = m.PotentiallyFixableIssuesInfo.suggestionEdit.Update(msg)
-	cmds = append(cmds, cmd)
+			return m, tea.Quit
+		case "esc":
+			return m, m.exitOrMoveToCssSelection()
+		}
+	case tea.WindowSizeMsg:
+		m.ready = true
+		m.height = msg.Height
+		m.width = msg.Width
 
-	m.PotentiallyFixableIssuesInfo.suggestionDisplay, cmd = m.PotentiallyFixableIssuesInfo.suggestionDisplay.Update(msg)
-	cmds = append(cmds, cmd)
+		// TODO: this logic needs to be handled, but I am not too sure how to properly handle this without knowing the exact way of determining the height while not necessarily knowing the size of the text
+		m.PotentiallyFixableIssuesInfo.scrollbar, cmd = m.PotentiallyFixableIssuesInfo.scrollbar.Update(tui.HeightMsg(m.PotentiallyFixableIssuesInfo.suggestionDisplay.Height))
+		cmds = append(cmds, cmd)
 
-	m.PotentiallyFixableIssuesInfo.scrollbar, cmd = m.PotentiallyFixableIssuesInfo.scrollbar.Update(m.PotentiallyFixableIssuesInfo.suggestionDisplay)
-	cmds = append(cmds, cmd)
+		cmds = append(cmds, tea.ClearScreen)
+	case error:
+		m.Err = msg
+		return m, tea.Quit
+	}
+
+	switch m.currentStage {
+	case sectionBreak:
+		cmd = m.handleSectionBreakMsgs(msg)
+		cmds = append(cmds, cmd)
+	case suggestionsProcessing:
+		cmd = m.handleSuggestionMsgs(msg)
+		cmds = append(cmds, cmd)
+	}
 
 	switch msg := msg.(type) {
 	case tea.KeyMsg:
 		key := msg.String()
 		switch m.currentStage {
-		case sectionBreak:
-			switch key {
-			case "enter":
-				*m.sectionBreakInfo.contextBreak = strings.TrimSpace(m.sectionBreakInfo.input.Value())
-				if *m.sectionBreakInfo.contextBreak != "" {
-					m.currentStage = suggestionsProcessing
-
-					cmd, err := m.setupForNextSuggestions()
-					if err != nil {
-						m.Err = err
-
-						return m, tea.Quit
-					}
-
-					cmds = append(cmds, cmd)
-				}
-			case "ctrl+c":
-				m.Err = ErrUserKilledProgram
-
-				return m, tea.Quit
-			case "esc":
-				return m, tea.Quit
-			}
-		case suggestionsProcessing:
-			if m.PotentiallyFixableIssuesInfo.isEditing {
-				switch key {
-				case "ctrl+c":
-					m.Err = ErrUserKilledProgram
-
-					return m, tea.Quit
-				case "esc":
-					return m, m.exitOrMoveToCssSelection()
-				case "ctrl+s":
-					m.PotentiallyFixableIssuesInfo.currentSuggestionState.currentSuggestion = alignWhitespace(m.PotentiallyFixableIssuesInfo.currentSuggestionState.original, m.PotentiallyFixableIssuesInfo.suggestionEdit.Value())
-					m.PotentiallyFixableIssuesInfo.isEditing = false
-
-					var err error
-					m.PotentiallyFixableIssuesInfo.currentSuggestionState.display, err = getStringDiff(m.PotentiallyFixableIssuesInfo.currentSuggestionState.original, m.PotentiallyFixableIssuesInfo.currentSuggestionState.currentSuggestion)
-					if err != nil {
-						m.Err = err
-
-						return m, tea.Quit
-					}
-
-					return m, m.setSuggestionDisplay()
-				case "ctrl+e":
-					m.PotentiallyFixableIssuesInfo.isEditing = false
-
-					var err error
-					m.PotentiallyFixableIssuesInfo.currentSuggestionState.display, err = getStringDiff(m.PotentiallyFixableIssuesInfo.currentSuggestionState.original, m.PotentiallyFixableIssuesInfo.currentSuggestionState.currentSuggestion)
-					if err != nil {
-						m.Err = err
-
-						return m, tea.Quit
-					}
-				case "ctrl+r":
-					m.PotentiallyFixableIssuesInfo.suggestionEdit.SetValue(m.PotentiallyFixableIssuesInfo.currentSuggestionState.originalSuggestion)
-				}
-			} else {
-				switch key {
-				case "ctrl+c":
-					m.Err = ErrUserKilledProgram
-
-					return m, tea.Quit
-				case "esc":
-					return m, m.exitOrMoveToCssSelection()
-				case "right":
-					cmd, err := m.moveToNextSuggestion()
-					if err != nil {
-						m.Err = err
-
-						return m, tea.Quit
-					}
-
-					cmds = append(cmds, cmd)
-				case "left":
-					cmd = m.moveToPreviousSuggestion()
-					cmds = append(cmds, cmd)
-				case "c":
-					// Copy original value to the clipboard
-					// original, err := repairUnicode(m.currentFileState.original)
-					// if err != nil {
-					// 	m.Err = err
-
-					// 	return m, tea.Quit
-					// }
-
-					// err = clipboard.WriteAll(original)
-					// TODO: make sure values are utf-8 compliant
-					err := clipboard.WriteAll(m.PotentiallyFixableIssuesInfo.currentSuggestionState.original)
-					if err != nil {
-						m.Err = err
-
-						return m, tea.Quit
-					}
-				case "enter":
-					if !m.PotentiallyFixableIssuesInfo.currentSuggestionState.isAccepted && m.PotentiallyFixableIssuesInfo.currentSuggestion != nil {
-						var replaceCount = 1
-						if m.PotentiallyFixableIssuesInfo.currentSuggestion.UpdateAllInstances {
-							replaceCount = -1
-						}
-
-						m.PotentiallyFixableIssuesInfo.FileTexts[m.PotentiallyFixableIssuesInfo.currentFileIndex] = strings.Replace(m.PotentiallyFixableIssuesInfo.FileTexts[m.PotentiallyFixableIssuesInfo.currentFileIndex], m.PotentiallyFixableIssuesInfo.currentSuggestionState.original, m.PotentiallyFixableIssuesInfo.currentSuggestionState.currentSuggestion, replaceCount)
-
-						m.PotentiallyFixableIssuesInfo.currentSuggestionState.isAccepted = true
-
-						if m.PotentiallyFixableIssuesInfo.currentSuggestion.AddCssSectionBreakIfMissing {
-							m.PotentiallyFixableIssuesInfo.AddCssSectionBreakIfMissing = true
-							m.PotentiallyFixableIssuesInfo.CssUpdateRequired = true
-						} else if m.PotentiallyFixableIssuesInfo.currentSuggestion.AddCssPageBreakIfMissing {
-							m.PotentiallyFixableIssuesInfo.AddCssPageBreakIfMissing = true
-							m.PotentiallyFixableIssuesInfo.CssUpdateRequired = true
-						}
-
-						cmd, err := m.moveToNextSuggestion()
-						if err != nil {
-							m.Err = err
-
-							return m, tea.Quit
-						}
-
-						cmds = append(cmds, cmd)
-					}
-				case "e":
-					if m.PotentiallyFixableIssuesInfo.currentSuggestionState != nil && !m.PotentiallyFixableIssuesInfo.currentSuggestionState.isAccepted {
-						m.PotentiallyFixableIssuesInfo.isEditing = true
-						m.PotentiallyFixableIssuesInfo.suggestionEdit.SetValue(m.PotentiallyFixableIssuesInfo.currentSuggestionState.currentSuggestion)
-
-						cmd = m.PotentiallyFixableIssuesInfo.suggestionEdit.Focus()
-						cmds = append(cmds, cmd)
-					}
-				}
-			}
 		case stageCssSelection:
 			switch key {
-			case "ctrl+c":
-				m.Err = ErrUserKilledProgram
-
-				return m, tea.Quit
 			case "up":
 				if m.CssSelectionInfo.currentCssIndex > 0 {
 					m.CssSelectionInfo.currentCssIndex--
@@ -318,39 +199,7 @@ func (m FixableIssuesModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 			case "enter":
 				m.currentStage = finalStage
 			}
-		default:
-			switch key {
-			case "ctrl+c":
-				m.Err = ErrUserKilledProgram
-
-				return m, tea.Quit
-			case "esc":
-				return m, tea.Quit
-			}
 		}
-	case tea.WindowSizeMsg:
-		m.ready = true
-		m.height = msg.Height
-		// TODO: see about removing this value...
-		maxDisplayHeight = msg.Height / 3
-		m.width = msg.Width
-
-		// TODO: replace with the logic for finding the body width and height
-		var maxWidth = m.width - columnPadding
-
-		m.PotentiallyFixableIssuesInfo.suggestionEdit.SetWidth(maxWidth)
-		m.PotentiallyFixableIssuesInfo.suggestionDisplay.Width = maxWidth - scrollbarPadding
-		if m.PotentiallyFixableIssuesInfo.suggestionDisplay.Height > maxDisplayHeight {
-			m.PotentiallyFixableIssuesInfo.suggestionDisplay.Height = maxDisplayHeight
-		}
-
-		m.PotentiallyFixableIssuesInfo.scrollbar, cmd = m.PotentiallyFixableIssuesInfo.scrollbar.Update(tui.HeightMsg(m.PotentiallyFixableIssuesInfo.suggestionDisplay.Height))
-		cmds = append(cmds, cmd)
-
-		cmds = append(cmds, tea.ClearScreen)
-	case error:
-		m.Err = msg
-		return m, tea.Quit
 	}
 
 	if m.currentStage == finalStage {
@@ -361,68 +210,6 @@ func (m FixableIssuesModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 	return m, tea.Batch(cmds...)
 }
-
-// func (m FixableIssuesModel) View() string {
-// 	var (
-// 		header = m.headerView()
-// 		s      strings.Builder
-// 	)
-
-// 	switch m.currentStage {
-// 	case sectionBreak:
-// 		s.WriteString("\n" + m.sectionBreakInfo.input.View() + "\n\n")
-// 	case suggestionsProcessing:
-// 		// s.WriteString(titleStyle.Render(fmt.Sprintf("Current File (%d/%d): %s ", m.PotentiallyFixableIssuesInfo.currentFileIndex+1, len(m.PotentiallyFixableIssuesInfo.FilePaths), m.PotentiallyFixableIssuesInfo.currentFile)) + "\n")
-// 		// s.WriteString(groupStyle.Render(fmt.Sprintf("Issue Group: %s", m.PotentiallyFixableIssuesInfo.currentSuggestionName) + "\n"))
-
-// 		if m.PotentiallyFixableIssuesInfo.currentSuggestionState == nil {
-// 			s.WriteString("\nNo current file is selected. Something may have gone wrong...\n\n")
-// 		} else {
-// 			var (
-// 				suggestion     = displayStyle.Width(m.width - columnPadding).Render(fmt.Sprintf(`"%s"`, m.PotentiallyFixableIssuesInfo.currentSuggestionState.display))
-// 				expectedHeight = strings.Count(suggestion, "\n") + 1
-// 			)
-// 			if m.PotentiallyFixableIssuesInfo.isEditing {
-// 				s.WriteString("\nEditing suggestion:\n\n")
-// 			} else if m.PotentiallyFixableIssuesInfo.currentSuggestionState.isAccepted {
-// 				s.WriteString("\n" + acceptedChangeTitleStyle.Render("Accepted change:") + "\n")
-// 			} else {
-// 				s.WriteString(fmt.Sprintf("\nSuggested change (%d/%d):\n", expectedHeight, maxDisplayHeight))
-// 			}
-
-// 			if m.PotentiallyFixableIssuesInfo.isEditing {
-// 				s.WriteString(suggestionBorderStyle.Render(m.PotentiallyFixableIssuesInfo.suggestionEdit.View()) + "\n\n")
-// 			} else {
-// 				if m.PotentiallyFixableIssuesInfo.suggestionDisplay.TotalLineCount() > m.PotentiallyFixableIssuesInfo.suggestionDisplay.VisibleLineCount() {
-// 					s.WriteString(suggestionBorderStyle.Render(lipgloss.JoinHorizontal(lipgloss.Top,
-// 						m.PotentiallyFixableIssuesInfo.suggestionDisplay.View(),
-// 						m.PotentiallyFixableIssuesInfo.scrollbar.View(),
-// 					)))
-// 				} else {
-// 					s.WriteString(suggestionBorderStyle.Render(m.PotentiallyFixableIssuesInfo.suggestionDisplay.View()))
-// 				}
-
-// 				s.WriteString(fmt.Sprintf("\033[0m\n\nSuggestion %d of %d.\n\n", m.PotentiallyFixableIssuesInfo.currentSuggestionIndex+1, len(m.PotentiallyFixableIssuesInfo.sectionSuggestionStates)))
-// 			}
-// 		}
-// 	case stageCssSelection:
-// 		s.WriteString("\nSelect the CSS file to modify:\n\n")
-// 		for i, cssFile := range m.cssSelectionInfo.cssFiles {
-// 			cursor := " "
-// 			if m.cssSelectionInfo.currentCssIndex == i {
-// 				cursor = ">"
-// 			}
-
-// 			s.WriteString(fmt.Sprintf("%s %d. %s\n", cursor, i+1, cssFile))
-// 		}
-
-// 		s.WriteString("\n")
-// 	}
-
-// 	m.displayControls(&s)
-
-// 	return header + s.String()
-// }
 
 func (m FixableIssuesModel) View() string {
 	if m.ready {
@@ -465,8 +252,11 @@ func (m FixableIssuesModel) getStageHeaders() string {
 }
 
 func (m FixableIssuesModel) footerView() string {
-	var s strings.Builder
-	s.WriteString(fillLine(controlsStyle.Render("Controls:"), m.width-footerBorderStyle.GetHorizontalBorderSize()) + "\n")
+	var (
+		s        strings.Builder
+		maxWidth = m.width - footerBorderStyle.GetHorizontalBorderSize()
+	)
+	s.WriteString(fillLine(controlsStyle.Render("Controls:"), maxWidth) + "\n")
 
 	var controls []string
 	switch m.currentStage {
@@ -511,8 +301,7 @@ func (m FixableIssuesModel) footerView() string {
 	}
 
 	var (
-		line     strings.Builder
-		maxWidth = m.width - columnPadding
+		line strings.Builder
 	)
 	for _, help := range controls {
 		if line.Len() == 0 {
@@ -530,236 +319,32 @@ func (m FixableIssuesModel) footerView() string {
 		}
 	}
 
-	// if line.Len() != 0 {
-	// 	s.WriteString("\n")
-	// }
-
 	return footerBorderStyle.Render(s.String())
 }
 
 func (m FixableIssuesModel) bodyView() string {
-	var s strings.Builder
-
 	switch m.currentStage {
 	case sectionBreak:
 		return m.sectionBreakView()
 	case suggestionsProcessing:
 		return m.suggestionsView()
-		// // s.WriteString(titleStyle.Render(fmt.Sprintf("Current File (%d/%d): %s ", m.PotentiallyFixableIssuesInfo.currentFileIndex+1, len(m.PotentiallyFixableIssuesInfo.FilePaths), m.PotentiallyFixableIssuesInfo.currentFile)) + "\n")
-		// // s.WriteString(groupStyle.Render(fmt.Sprintf("Issue Group: %s", m.PotentiallyFixableIssuesInfo.currentSuggestionName) + "\n"))
-
-		// if m.PotentiallyFixableIssuesInfo.currentSuggestionState == nil {
-		// 	s.WriteString("\nNo current file is selected. Something may have gone wrong...\n\n")
-		// } else {
-		// 	var (
-		// 		suggestion     = displayStyle.Width(m.width - columnPadding).Render(fmt.Sprintf(`"%s"`, m.PotentiallyFixableIssuesInfo.currentSuggestionState.display))
-		// 		expectedHeight = strings.Count(suggestion, "\n") + 1
-		// 	)
-		// 	if m.PotentiallyFixableIssuesInfo.isEditing {
-		// 		s.WriteString("\nEditing suggestion:\n\n")
-		// 	} else if m.PotentiallyFixableIssuesInfo.currentSuggestionState.isAccepted {
-		// 		s.WriteString("\n" + acceptedChangeTitleStyle.Render("Accepted change:") + "\n")
-		// 	} else {
-		// 		s.WriteString(fmt.Sprintf("\nSuggested change (%d/%d):\n", expectedHeight, maxDisplayHeight))
-		// 	}
-
-		// 	if m.PotentiallyFixableIssuesInfo.isEditing {
-		// 		s.WriteString(suggestionBorderStyle.Render(m.PotentiallyFixableIssuesInfo.suggestionEdit.View()) + "\n\n")
-		// 	} else {
-		// 		if m.PotentiallyFixableIssuesInfo.suggestionDisplay.TotalLineCount() > m.PotentiallyFixableIssuesInfo.suggestionDisplay.VisibleLineCount() {
-		// 			s.WriteString(suggestionBorderStyle.Render(lipgloss.JoinHorizontal(lipgloss.Top,
-		// 				m.PotentiallyFixableIssuesInfo.suggestionDisplay.View(),
-		// 				m.PotentiallyFixableIssuesInfo.scrollbar.View(),
-		// 			)))
-		// 		} else {
-		// 			s.WriteString(suggestionBorderStyle.Render(m.PotentiallyFixableIssuesInfo.suggestionDisplay.View()))
-		// 		}
-
-		// 		s.WriteString(fmt.Sprintf("\033[0m\n\nSuggestion %d of %d.\n\n", m.PotentiallyFixableIssuesInfo.currentSuggestionIndex+1, len(m.PotentiallyFixableIssuesInfo.sectionSuggestionStates)))
-		// 	}
-		// }
 	case stageCssSelection:
-		s.WriteString("\nSelect the CSS file to modify:\n\n")
-		for i, cssFile := range m.CssSelectionInfo.cssFiles {
-			cursor := " "
-			if m.CssSelectionInfo.currentCssIndex == i {
-				cursor = ">"
-			}
-
-			s.WriteString(fmt.Sprintf("%s %d. %s\n", cursor, i+1, cssFile))
-		}
-
-		s.WriteString("\n")
+		return m.cssSelectionView()
 	}
 
-	return s.String()
-}
-
-func (m *FixableIssuesModel) setupForNextSuggestions() (tea.Cmd, error) {
-	if m.logFile != nil {
-		fmt.Fprintln(m.logFile, "Getting next suggestions")
-	}
-
-	for m.PotentiallyFixableIssuesInfo.currentFileIndex < len(m.PotentiallyFixableIssuesInfo.FilePaths) {
-		m.PotentiallyFixableIssuesInfo.currentFile = m.PotentiallyFixableIssuesInfo.FilePaths[m.PotentiallyFixableIssuesInfo.currentFileIndex]
-		if m.logFile != nil {
-			fmt.Fprintf(m.logFile, "Current file is %q is %d of %d\n", m.PotentiallyFixableIssuesInfo.currentFile, m.PotentiallyFixableIssuesInfo.currentFileIndex+1, len(m.PotentiallyFixableIssuesInfo.FilePaths))
-		}
-
-		for m.PotentiallyFixableIssuesInfo.potentialFixableIssueIndex < len(m.PotentiallyFixableIssuesInfo.suggestions) {
-			var potentialFixableIssue = m.PotentiallyFixableIssuesInfo.suggestions[m.PotentiallyFixableIssuesInfo.potentialFixableIssueIndex]
-			if m.logFile != nil {
-				fmt.Fprintf(m.logFile, "Possible fixable issue %q is %d of %d issues.\n", potentialFixableIssue.Name, m.PotentiallyFixableIssuesInfo.potentialFixableIssueIndex+1, len(m.PotentiallyFixableIssuesInfo.suggestions))
-			}
-
-			if !m.runAll && (potentialFixableIssue.IsEnabled == nil || *potentialFixableIssue.IsEnabled) {
-				if m.logFile != nil {
-					fmt.Fprintf(m.logFile, "Skipping possible fixable issue %q with isEnabled set to %v\n", potentialFixableIssue.Name, potentialFixableIssue.IsEnabled)
-				}
-
-				m.PotentiallyFixableIssuesInfo.potentialFixableIssueIndex++
-				continue
-			}
-
-			var (
-				suggestions = potentialFixableIssue.GetSuggestions(m.PotentiallyFixableIssuesInfo.FileTexts[m.PotentiallyFixableIssuesInfo.currentFileIndex])
-			)
-
-			if m.logFile != nil {
-				fmt.Fprintf(m.logFile, "Possible fixable issue %q has %d suggestion(s) found\n", potentialFixableIssue.Name, len(suggestions))
-			}
-
-			if len(suggestions) != 0 {
-				m.PotentiallyFixableIssuesInfo.currentSuggestion = &potentialFixableIssue
-				m.PotentiallyFixableIssuesInfo.sectionSuggestionStates = make([]suggestionState, len(suggestions))
-
-				var i = 0
-				for original, suggestion := range suggestions {
-					var display, err = getStringDiff(original, suggestion)
-					if err != nil {
-						return nil, err
-					}
-
-					m.PotentiallyFixableIssuesInfo.sectionSuggestionStates[i] = suggestionState{
-						original:           original,
-						originalSuggestion: suggestion,
-						currentSuggestion:  suggestion,
-						display:            display,
-					}
-
-					i++
-				}
-
-				m.PotentiallyFixableIssuesInfo.currentSuggestionIndex = 0
-				m.PotentiallyFixableIssuesInfo.currentSuggestionState = &m.PotentiallyFixableIssuesInfo.sectionSuggestionStates[0]
-				cmd := m.setSuggestionDisplay()
-				m.PotentiallyFixableIssuesInfo.currentSuggestionName = potentialFixableIssue.Name
-
-				m.PotentiallyFixableIssuesInfo.potentialFixableIssueIndex++
-
-				return cmd, nil
-			}
-
-			m.PotentiallyFixableIssuesInfo.potentialFixableIssueIndex++
-		}
-
-		m.PotentiallyFixableIssuesInfo.currentFileIndex++
-		m.PotentiallyFixableIssuesInfo.potentialFixableIssueIndex = 0
-	}
-
-	if m.PotentiallyFixableIssuesInfo.CssUpdateRequired {
-		m.currentStage = stageCssSelection
-	} else {
-		m.currentStage = finalStage
-	}
-
-	return nil, nil
-}
-
-func (m *FixableIssuesModel) moveToNextSuggestion() (tea.Cmd, error) {
-	if m.PotentiallyFixableIssuesInfo.currentSuggestionIndex+1 < len(m.PotentiallyFixableIssuesInfo.sectionSuggestionStates) {
-		m.PotentiallyFixableIssuesInfo.currentSuggestionIndex++
-		m.PotentiallyFixableIssuesInfo.currentSuggestionState = &m.PotentiallyFixableIssuesInfo.sectionSuggestionStates[m.PotentiallyFixableIssuesInfo.currentSuggestionIndex]
-
-		return m.setSuggestionDisplay(), nil
-	}
-
-	return m.setupForNextSuggestions()
-}
-
-func (m *FixableIssuesModel) moveToPreviousSuggestion() tea.Cmd {
-	if m.PotentiallyFixableIssuesInfo.currentSuggestionIndex > 0 {
-		m.PotentiallyFixableIssuesInfo.currentSuggestionIndex--
-		m.PotentiallyFixableIssuesInfo.currentSuggestionState = &m.PotentiallyFixableIssuesInfo.sectionSuggestionStates[m.PotentiallyFixableIssuesInfo.currentSuggestionIndex]
-		return m.setSuggestionDisplay()
-	}
-
-	return nil
-}
-
-func (m *FixableIssuesModel) setSuggestionDisplay() tea.Cmd {
-	if m.PotentiallyFixableIssuesInfo.currentSuggestionState == nil {
-		return nil
-	}
-
-	if m.logFile != nil {
-		fmt.Fprintf(m.logFile, "current width %d; border width: %d; scrollbar padding: %d\n", m.PotentiallyFixableIssuesInfo.suggestionDisplay.Width, suggestionBorderStyle.GetHorizontalBorderSize(), scrollbarPadding)
-	}
-
-	var (
-		expectedSuggestionWidth = m.PotentiallyFixableIssuesInfo.suggestionDisplay.Width
-		suggestion              = displayStyle.Render(wrapLines(fmt.Sprintf(`"%s"`, m.PotentiallyFixableIssuesInfo.currentSuggestionState.display), expectedSuggestionWidth))
-	)
-
-	if m.logFile != nil {
-		fmt.Fprintf(m.logFile, "New suggestion getting set with width %d and a value of %q\n", expectedSuggestionWidth, suggestion)
-	}
-
-	m.PotentiallyFixableIssuesInfo.suggestionDisplay.SetContent(suggestion)
-
-	// TODO: how do I handle the resizing of the suggestion display
-	// one option would be to
-	var cmd tea.Cmd
-	m.PotentiallyFixableIssuesInfo.scrollbar, cmd = m.PotentiallyFixableIssuesInfo.scrollbar.Update(tui.HeightMsg(m.PotentiallyFixableIssuesInfo.suggestionDisplay.Height))
-
-	return cmd
+	return ""
 }
 
 func (m *FixableIssuesModel) exitOrMoveToCssSelection() tea.Cmd {
 	if m.PotentiallyFixableIssuesInfo.CssUpdateRequired {
 		m.currentStage = stageCssSelection
+
+		if len(m.CssSelectionInfo.cssFiles) != 0 {
+			m.CssSelectionInfo.SelectedCssFile = m.CssSelectionInfo.cssFiles[m.CssSelectionInfo.currentCssIndex]
+		}
 	} else {
 		return tea.Quit
 	}
 
 	return nil
-}
-
-func getStringDiff(original, new string) (string, error) {
-	return stringdiff.GetPrettyDiffString(strings.TrimLeft(original, "\n"), strings.TrimLeft(new, "\n"))
-}
-
-// textarea gets rid of tabs when creating changes, so in order to preserve tabs in the starting whitespace of a line
-// we will use the value of original as the template for what whitespace is needed for each line present
-func alignWhitespace(original, new string) string {
-	origLines := strings.Split(original, "\n")
-	newLines := strings.Split(new, "\n")
-
-	var min = len(newLines)
-	if len(origLines) < min {
-		min = len(origLines)
-	}
-
-	for i := 0; i < min; i++ {
-		origPrefix := ""
-		for j := 0; j < len(origLines[i]); j++ {
-			if !unicode.IsSpace(rune(origLines[i][j])) {
-				break
-			}
-			origPrefix += string(origLines[i][j])
-		}
-		newLines[i] = origPrefix + strings.TrimLeft(newLines[i], " \t")
-	}
-
-	return strings.Join(newLines, "\n")
 }
